orders/src/routes/delete.ts
import express, { Request, Response } from 'express';
import { Order, OrderStatus } from '../models/order';
import { NotAuthorizedError, NotFoundError, requireAuth, validateRequest } from '@shogglearningtixcommon/common';
import { param } from 'express-validator';
import { OrderCancelledPublisher } from '../events/publishers/order-cancelled-publisher';
import { natsWrapper } from '../nats-wrapper';

const router = express.Router();

router.delete(
  '/api/orders/:orderId',
  requireAuth,
  [
    // If the following validation checks fail, they append properties onto the request object
    param('orderId').isMongoId().withMessage('Order Id must be valid Id'),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const { orderId } = req.params;

    const order = await Order.findById(orderId).populate('ticket');

    if(!order) {
      throw new NotFoundError();
    }

    if(order.userId !== req.currentUser!.id) {
      throw new NotAuthorizedError();
    }

    order.status = OrderStatus.Cancelled;
    await order.save();

    // Publish an event saying that an order was cancelled
    new OrderCancelledPublisher(natsWrapper.client).publish({
      id: order.id,
      version: order.version,
      ticket: {
        id: order.ticket.id
      }
    });

    // TODO this should really be a patch as we shouldnt be returning a body in a 204.
    res.status(204).send(order);
  }
);

export { router as deleteOrderRouter };